Consolidate sandbox volume tests into a single comprehensive test file
